Move diffusion test utils to separate file
