+++ model/atmosphere/diffusion/tests/diffusion_tests/mpi_tests/test_parallel_diffusion.py
@@ -9,6 +9,7 @@
 import pytest
 
 import icon4py.model.common.test_utils.datatest_utils
+import icon4py.model.common.test_utils.diffusion_utils
 from icon4py.model.atmosphere.diffusion import diffusion as diffusion_
 from icon4py.model.common import dimension as dims
 from icon4py.model.common.decomposition import definitions
@@ -59,13 +60,13 @@ def test_parallel_diffusion(
     print(
         f"rank={processor_props.rank}/{processor_props.comm_size}: using local grid with {icon_grid.num_cells} Cells, {icon_grid.num_edges} Edges, {icon_grid.num_vertices} Vertices"
     )
-    metric_state = icon4py.model.common.test_utils.datatest_utils.construct_metric_state(
+    metric_state = icon4py.model.common.test_utils.diffusion_utils.construct_metric_state(
         metrics_savepoint
     )
     cell_geometry = grid_savepoint.construct_cell_geometry()
     edge_geometry = grid_savepoint.construct_edge_geometry()
     interpolation_state = (
-        icon4py.model.common.test_utils.datatest_utils.construct_interpolation_state(
+        icon4py.model.common.test_utils.diffusion_utils.construct_interpolation_state(
             interpolation_savepoint
         )
     )
@@ -76,7 +77,7 @@ def test_parallel_diffusion(
         stretch_factor=stretch_factor,
         rayleigh_damping_height=damping_height,
     )
-    config = icon4py.model.common.test_utils.datatest_utils.construct_config(
+    config = icon4py.model.common.test_utils.diffusion_utils.construct_config(
         experiment, ndyn_substeps=ndyn_substeps
     )
     diffusion_params = diffusion_.DiffusionParams(config)
@@ -101,7 +102,7 @@ def test_parallel_diffusion(
         cell_params=cell_geometry,
     )
     print(f"rank={processor_props.rank}/{processor_props.comm_size}: diffusion initialized ")
-    diagnostic_state = icon4py.model.common.test_utils.datatest_utils.construct_diagnostics(
+    diagnostic_state = icon4py.model.common.test_utils.diffusion_utils.construct_diagnostics(
         diffusion_savepoint_init
     )
     prognostic_state = diffusion_savepoint_init.construct_prognostics()
