+++ model/atmosphere/diffusion/tests/diffusion_tests/test_diffusion.py
@@ -19,7 +19,7 @@ from icon4py.model.common.test_utils import (
     reference_funcs as ref_funcs,
     serialbox_utils as sb,
 )
-from icon4py.model.common.test_utils.datatest_utils import (
+from icon4py.model.common.test_utils.diffusion_utils import (
     construct_config,
     construct_diagnostics,
     construct_interpolation_state,
